# docs/swagger.yaml
basePath: /
definitions:
  web.CreateRoomRequest:
    properties:
      name:
        type: string
      ownerID:
        type: string
      type:
        enum:
        - public
        - private
        type: string
    required:
    - name
    - type
    type: object
  web.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  web.RegisterRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: Documentation API for Real Time Chat using Fiber + WebSocket
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Real Time Chat API + WebSocket
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login request
        in: body
        name: LoginRequest
        required: true
        schema:
          $ref: '#/definitions/web.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user to the system
      parameters:
      - description: Register request
        in: body
        name: RegisterRequest
        required: true
        schema:
          $ref: '#/definitions/web.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Register new user
      tags:
      - auth
  /rooms:
    get:
      description: Get all available chat rooms
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List all rooms
      tags:
      - rooms
    post:
      consumes:
      - application/json
      description: Create a new chat room
      parameters:
      - description: Create room request
        in: body
        name: CreateRoomRequest
        required: true
        schema:
          $ref: '#/definitions/web.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create new room
      tags:
      - rooms
  /rooms/{id}/message:
    get:
      description: Get messages from a specific room with pagination
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Messages per page (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get room messages
      tags:
      - rooms
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" and JWT token value of the logged in user to access
      this API.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"